Add README, refactor and lint
