+++ test/test-ingredients.js
@@ -1,0 +1,125 @@
+let testIngredients = [
+  {
+    "id": 20081,
+    "name": "wheat flour",
+    "estimatedCostInCents": 142
+  },
+  {
+    "id": 18372,
+    "name": "bicarbonate of soda",
+    "estimatedCostInCents": 582
+  },
+  {
+    "id": 1123,
+    "name": "eggs",
+    "estimatedCostInCents": 472
+  },
+  {
+    "id": 19335,
+    "name": "sucrose",
+    "estimatedCostInCents": 902
+  },
+  {
+    "id": 19334,
+    "name": "brown sugar",
+    "estimatedCostInCents": 559
+  },
+  {
+    "id": 1009016,
+    "name": "apple cider",
+    "estimatedCostInCents": 468
+  },
+  {
+    "id": 9003,
+    "name": "apple",
+    "estimatedCostInCents": 207
+  },
+  {
+    "id": 20027,
+    "name": "corn starch",
+    "estimatedCostInCents": 236
+  },
+  {
+    "id": 1002046,
+    "name": "dijon style mustard",
+    "estimatedCostInCents": 619
+  },
+  {
+    "id": 11215,
+    "name": "whole garlic clove",
+    "estimatedCostInCents": 220
+  },
+  {
+    "id": 1012046,
+    "name": "whole grain dijon mustard",
+    "estimatedCostInCents": 867
+  },
+  {
+    "id": 1002030,
+    "name": "black pepper",
+    "estimatedCostInCents": 441
+  },
+  {
+    "id": 1001,
+    "name": "butter",
+    "estimatedCostInCents": 618
+  },
+  {
+    "id": 4582,
+    "name": "oil",
+    "estimatedCostInCents": 807
+  },
+  {
+    "id": 2009,
+    "name": "red chili powder",
+    "estimatedCostInCents": 499
+  },
+  {
+    "id": 6168,
+    "name": "tabasco sauce",
+    "estimatedCostInCents": 859
+  },
+  {
+    "id": 18371,
+    "name": "baking powder",
+    "estimatedCostInCents": 346
+  },
+  {
+    "id": 9040,
+    "name": "ripe banana",
+    "estimatedCostInCents": 331
+  },
+  {
+    "id": 20011,
+    "name": "buck wheat flour",
+    "estimatedCostInCents": 865
+  },
+  {
+    "id": 1230,
+    "name": "buttermilk",
+    "estimatedCostInCents": 773
+  },
+
+  {
+    "id": 2048,
+    "name": "apple cider vinegar",
+    "estimatedCostInCents": 532
+  },
+  {
+    "id": 20090,
+    "name": "brown rice flour",
+    "estimatedCostInCents": 667
+  },
+  {
+    "id": 93784,
+    "name": "brown rice syrup",
+    "estimatedCostInCents": 126
+  },
+  {
+    "id": 1124,
+    "name": "egg albumen",
+    "estimatedCostInCents": 304
+  }
+]
+
+module.exports = testIngredients;
