+++ test/test-recipes.js
@@ -1,0 +1,314 @@
+const testRecipes = [
+  {
+    "id": 595736,
+    "image": "https://spoonacular.com/recipeImages/595736-556x370.jpg",
+    "ingredients": [
+      {
+        "id": 20081,
+        "quantity": {
+          "amount": 1.5,
+          "unit": "c"
+        }
+      },
+      {
+        "id": 18372,
+        "quantity": {
+          "amount": 0.5,
+          "unit": "tsp"
+        }
+      },
+      {
+        "id": 1123,
+        "quantity": {
+          "amount": 1,
+          "unit": "large"
+        }
+      },
+      {
+        "id": 19335,
+        "quantity": {
+          "amount": 0.5,
+          "unit": "c"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "In a large mixing bowl, whisk together the dry ingredients (flour, pudding mix, soda and salt). Set aside.In a large mixing bowl of a stand mixer, cream butter for 30 seconds. Gradually add granulated sugar and brown sugar and cream until light and fluffy.",
+        "number": 1
+      },
+      {
+        "instruction": "Add egg and vanilla and mix until combined.",
+        "number": 2
+      },
+      {
+        "instruction": "Add dry ingredients and mix on low just until incorporated. Stir in chocolate chips.Scoop the dough into 1,5 tablespoon size balls and place on a plate or sheet. Cover with saran wrap and chill at least 2 hours or overnight.When ready to bake, preheat oven to 350 degrees.",
+        "number": 3
+      }
+    ],
+    "name": "Dijon Loaded Chocolate Chip Pudding Cookie Cups",
+    "tags": [
+      "testSingle",
+      "testMultiple1",
+      "testMultiple2",
+      "appetizer",
+      "antipasto",
+      "hor d'oeuvre"
+    ]
+  },
+  {
+    "id": 678353,
+    "image": "https://spoonacular.com/recipeImages/678353-556x370.jpg",
+    "ingredients": [
+      {
+        "id": 1009016,
+        "quantity": {
+          "amount": 1.5,
+          "unit": "cups"
+        }
+      },
+      {
+        "id": 9003,
+        "quantity": {
+          "amount": 2,
+          "unit": ""
+        }
+      },
+      {
+        "id": 20027,
+        "quantity": {
+          "amount": 1,
+          "unit": "tablespoon"
+        }
+      },
+      {
+        "id": 1002046,
+        "quantity": {
+          "amount": 1,
+          "unit": "tablespoon"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "Season the pork chops with salt and pepper and grill or pan fry over medium high heat until cooked, about 3-5 minutes per side. (If grilling, baste the chops in the maple dijon apple cider sauce as you grill.)Meanwhile, mix the remaining ingredients except the apple slices, bring to a simmer and cook until the sauce thickens, about 2-5 minutes.Grill or saute the apple slices until just tender but still crisp.Toss the pork chops and apple slices in the maple dijon apple cider sauce and enjoy!",
+        "number": 1
+      }
+    ],
+    "name": "Maple Dijon Apple Cider Grilled Pork Chops",
+    "tags": [
+      "lunch",
+      "testMultiple2",
+      "dinner"
+    ]
+  },
+  {
+    "id": 412309,
+    "image": "https://spoonacular.com/recipeImages/412309-556x370.jpeg",
+    "ingredients": [
+      {
+        "id": 1002030,
+        "quantity": {
+          "amount": 4,
+          "unit": "teaspoons"
+        }
+      },
+      {
+        "id": 19334,
+        "quantity": {
+          "amount": 8,
+          "unit": "tablespoons"
+        }
+      },
+      {
+        "id": 1001,
+        "quantity": {
+          "amount": 2,
+          "unit": "cups"
+        }
+      },
+      {
+        "id": 4582,
+        "quantity": {
+          "amount": 4,
+          "unit": "servings"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "Mix the hot sauce, butter, mango habanero sauce, brown sugar, chili powder, garlic powder, onion powder, black pepper, cayenne pepper and seasoning salt in a bowl. Stir vigorously until completely combined.",
+        "number": 1
+      }
+    ],
+    "name": "Dirty Steve's Original Wing Sauce",
+    "tags": [
+      "testSingle"
+    ]
+  },
+  {
+    "id": 741603,
+    "image": "https://spoonacular.com/recipeImages/741603-556x370.jpeg",
+    "ingredients": [
+      {
+        "id": 20081,
+        "quantity": {
+          "amount": 1,
+          "unit": "cup"
+        }
+      },
+      {
+        "id": 18371,
+        "quantity": {
+          "amount": 2,
+          "unit": "teaspoons"
+        }
+      },
+      {
+        "id": 9040,
+        "quantity": {
+          "amount": 12,
+          "unit": "servings"
+        }
+      },
+      {
+        "id": 20011,
+        "quantity": {
+          "amount": 1,
+          "unit": "cup"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "Watch how to make this recipe.",
+        "number": 1
+      },
+      {
+        "instruction": "In a large bowl, whisk together buttermilk, eggs, baking powder, sugar, salt and butter.",
+        "number": 2
+      },
+      {
+        "instruction": "In another large bowl mix together all-purpose flour and buckwheat flour.",
+        "number": 3
+      }
+    ],
+    "name": "Elvis Pancakes",
+    "tags": [
+      "testSingle"
+    ]
+  },
+  {
+    "id": 562334,
+    "image": "https://spoonacular.com/recipeImages/562334-556x370.jpg",
+    "ingredients": [
+      {
+        "id": 2048,
+        "quantity": {
+          "amount": 2,
+          "unit": "teaspoons"
+        }
+      },
+      {
+        "id": 18371,
+        "quantity": {
+          "amount": 0.3333333333333333,
+          "unit": "teaspoon"
+        }
+      },
+      {
+        "id": 20090,
+        "quantity": {
+          "amount": 1.125,
+          "unit": "cup"
+        }
+      },
+      {
+        "id": 93784,
+        "quantity": {
+          "amount": 2,
+          "unit": "teaspoons"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "Grease or spray oil a 9\u00d75 inch loaf pan.Preheat oven to 170 \u2013 200\u00b0F (lowest possible).",
+        "number": 1
+      },
+      {
+        "instruction": "Mix warm water with brown rice syrup, molasses, and yeast in a cup larger than 8 oz., as it may bubble over; set aside until foamy on the top, no more than 5 minutes.In the bowl of your mixer, beat the eggs at high speed in a large mixing bowl until large bubbles form, about 20 seconds.",
+        "number": 2
+      },
+      {
+        "instruction": "Whisk together the dry ingredients; set aside.",
+        "number": 3
+      }
+    ],
+    "name": "Name Search Test",
+    "tags": []
+  },
+  {
+    "id": 507921,
+    "image": "https://spoonacular.com/recipeImages/507921-556x370.jpg",
+    "ingredients": [
+      {
+        "id": 18371,
+        "quantity": {
+          "amount": 1,
+          "unit": "teaspoon"
+        }
+      },
+      {
+        "id": 19350,
+        "quantity": {
+          "amount": 0.25,
+          "unit": "cup"
+        }
+      },
+      {
+        "id": 1123,
+        "quantity": {
+          "amount": 1,
+          "unit": ""
+        }
+      },
+      {
+        "id": 1124,
+        "quantity": {
+          "amount": 2,
+          "unit": "large"
+        }
+      },
+      {
+        "id": 20081,
+        "quantity": {
+          "amount": 1,
+          "unit": "cup"
+        }
+      }
+    ],
+    "instructions": [
+      {
+        "instruction": "To make the Cupcakes: Preheat oven to 350 degrees. Line 12 cupcake tins with paper holders.",
+        "number": 1
+      },
+      {
+        "instruction": "Whisk together dry Fruit Cocktail Cupcakes ingredients.",
+        "number": 2
+      },
+      {
+        "instruction": "Add in wet Fruit Cocktail Cupcakes ingredients and stir with a rubber spatula until thoroughly combined. Fill cupcake tins evenly, and bake for 20 minutes or until thin knife inserted in center comes out clean.",
+        "number": 3
+      }
+    ],
+    "name": "3 Tag Search Test",
+    "tags": [
+      "testMultiple1",
+      "testMultiple2",
+      "testMultiple3"
+    ]
+  }
+]
+
+module.exports = testRecipes;
