--- tests/test-user.js
@@ -1,25 +1,0 @@
-let usersData = [
-  {
-    "name": "Saige O'Kon",
-    "id": 1,
-    "pantry": [
-      {
-        "ingredient": 11477,
-        "amount": 4
-      },
-      {
-        "ingredient": 11297,
-        "amount": 4
-      },
-      {
-        "ingredient": 1082047,
-        "amount": 10
-      },
-      {
-        "ingredient": 20081,
-        "amount": 5
-      }
-    ]
-  }
-]
-  module.exports = usersData;
